Allow the HTTPD server to work in a proxy mode. If you define a custom path pointing to http://... or https://... it will internally redirect the request to the remote server, and serve the response back to the user from the local server.
For example:

cordova.plugins.CorHttpd.startServer({
                    'www_root': '',
                    'port': 8080,
                    'localhost_only': false,
                    'custom_paths': {
                        '/imgur/': 'http://i.imgur.com/'
                    }
                }, function(){console.log("SUCCESS");}, function(){console.log("FAILURE");});

When you then fetch http://localhost:8080/imgur/TNZYux0.jpg, you will fetch the content from imgur, and serve it as if it came from your own domain.

The reason for doing this... CrossDomain issues on image content served from an off device domain, and not wanting to apply unrestricted CORS access to all domains.

NOTE:
On iOS, this will download the entire contents of the response into memory before sending it on to the calling client. On Android, this is chunked in up to 16kb segments. Someone with more iOS experience than myself can probably make the iOS side of this behave more efficiently.